Adaptar rodaCruzada.sh para rodar a partir de uma fold N

+++ run/rodaCruzada.sh
@@ -6,7 +6,7 @@
 export CUDA_VISIBLE_DEVICES=0
 
 # IMPORTANTE: 3 dobras é muito pouco. Usei apenas para rodar mais apidamente um exemplo.
-ndobras=5  
+ndobras=5
 rodaPadrao=true
 rodaSiamesa=true
 
@@ -29,12 +29,64 @@ pastaTreino="../data/train"
 pastaTeste="../data/test"
 pastaResultados="../results"
 pastaDobrasResultados="../resultsNfolds"
+# Local do arquivo de dobras completas
+dobrasCompletasArq="./dobrasCompletasArq.txt"
 
 folds=()
 for((i=1;i<=$ndobras;i+=1)); do folds+=("fold_${i}"); done
 
-mkdir -p ../results_dl/           
-rm -rf ../results_dl/*
+# Cria pastas de resultados
+mkdir -p ../results_dl/
+mkdir -p ${pastaResultados}
+mkdir -p ${pastaTreino} 
+mkdir -p ${pastaTeste}
+mkdir -p ${pastaDobrasResultados}
+
+# Cria dobrasCompletasArq caso não exista
+if [ ! -f $dobrasCompletasArq ]; then
+   touch $dobrasCompletasArq
+fi
+
+# Coloca as dobras existentes em um array
+dobrasCompletas=($(cat $dobrasCompletasArq))
+
+# Registra dobras restantes
+dobrasRestantes=()
+for fold in "${folds[@]}"; do
+   if [[ ! " ${dobrasCompletas[@]} " =~ " ${fold} " ]]; then
+      dobrasRestantes+=("${fold}")
+   fi
+done
+
+# Confere se o usuário quer continuar os testes
+if [ ${#dobrasCompletas[@]} -gt 0 ]; then
+   echo "${#dobrasCompletas[@]} testes de ${ndobras} executados."
+   read -p "Deseja continuar? (s/n):" choice
+   if [[ "$choice" == "n" || "$choice" == "N" ]]; then
+      # Apagar dobras completadas e reiniciar
+      echo "Reiniciando testes..."
+      rm -rf ${pastaTreino}/* ${pastaTeste}/* ${pastaResultados}/* ${pastaDobrasResultados}/*
+      rm -rf ../results_dl/*
+      echo  'run,learning_rate,architecture,optimizer,precision,recall,fscore' > ../results_dl/results.csv
+      rm $dobrasCompletasArq
+      touch $dobrasCompletasArq
+      # Reinicia dobrasRestantes e Completas
+      dobrasRestantes=("${folds[@]}")
+      dobrasCompletas=()
+   else
+      echo "Continuando testes nas dobras restantes..."
+      # Atualiza dobrasCompletas e reinicia
+      dobrasCompletas=($(cat "$dobrasCompletasArq"))
+      dobrasRestantes=()
+      for fold in "${folds[@]}"; do
+         if [[ ! " ${dobrasCompletas[@]} " =~ " ${fold}" ]]; then
+            dobrasRestantes+=("${fold}")
+         fi
+      done
+   fi
+else
+   echo "Iniciando novo teste..."
+fi
 
 if [ "$procedimento" != "teste" ]
 then
@@ -42,19 +94,10 @@ then
    rm -rf ../model_checkpoints/*
 fi
 
-echo  'run,learning_rate,architecture,optimizer,precision,recall,fscore' > ../results_dl/results.csv
-
-mkdir -p ${pastaResultados}
-mkdir -p ${pastaTreino} 
-mkdir -p ${pastaTeste}
-
-rm -rf ${pastaDobrasResultados}/*
-mkdir -p ${pastaDobrasResultados}
-
 #Mudei este log de lugar, esta junto com os .output agora
 #rm /tmp/deep_learning*log*
 
-for Teste in "${folds[@]}"
+for Teste in "${dobrasRestantes[@]}"
 do
   
    echo 'Preparing test on' ${Teste} '...'
@@ -71,8 +114,6 @@ do
          echo 'Adding to train' ${outro} 
          cp -R ${pastaDobrasImagens}/${outro}/* ${pastaTreino} 
       fi   
-
-
    done
    
    run=${Teste#*_}
@@ -86,9 +127,9 @@ do
   
    mkdir -p ${pastaDobrasResultados}/${Teste}
    mv ${pastaResultados}/* ${pastaDobrasResultados}/${Teste}   
-    
+   
+   echo $Teste >> $dobrasCompletasArq
    #break
-
 done
 
 if [ "$procedimento" != "treino" ]
